Add post action to create new board members during on-boarding This action allows the org creator to add board members to their organization. Removed the params for email since I don't think the org creator will necessarily want to put all the board member emails in at the moment (also they might not have them ready). In the User model you will see that I had to add an if statement to the email validation and comment out the has_secure_password for now so that we can allow the creation of very basic User instances of type, "Board".
After a new user is created, the person will click "add new member" which triggers the form submission and post action. The post action always redirects back to /users/new. When the user is done creating their board members, they can click "continue" which is just a link to /meetings (will be a list of their org's meetings).