Add basic has_many-belongs_to association for user and meeting_attendees
This allows this AR method call to work: meeting.meeting_attendees.first.user. It returns a User object.

Note: what we might want instead is to be able to go meeting.meeting_attendees and then that returns a list of User objects...will need to explore this more complex association. It may be similar to the way responsible_users is set up, or chairs/movers/seconders. May also be similar to meeting_permissions?

Meeting has_many MeetingAttendees through Users? or vice-versa?